requirements.txt
# Location: /techcareer-analyzer/requirements.txt

# Core ML/Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4

# Deep Learning
torch==2.1.1
transformers==4.36.2
tensorflow==2.15.0
keras==2.15.0

# NLP
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2

# Data Collection
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.16.0
scrapy==2.11.0

# Web Framework
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Time Series
prophet==1.1.5
statsmodels==0.14.1

# API Integration
linkedin-api==2.2.0
python-dotenv==1.0.0

# Utilities
joblib==1.3.2
tqdm==4.66.1
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0